[Version] App: 2.1.2 beta 3 Database: V5 [Features] Add enableLosePoints mode && multipleChoicesQuestionsCheckingStrategy in GradingSetting [Fix] DynamicValidator still throws exception when ignoreMissingField = true && the field is missing in some cases Wrong readonly display("Lock" icon) in QuestionEditor.vue User submissions and scores missing in ExamRecordDetailView.vue when the user submits null answer Wrong "change question author" & "change question group author" permission request when update questions UpdateQuestionsAction sending unuseful messages [Improve] Auto logout

+++ src/main/java/indi/etern/checkIn/entities/question/interfaces/answer/Answer.java
@@ -7,7 +7,7 @@ public abstract class Answer<QuestionType,Source> {
     public enum CheckedResultType {
         CORRECT,HALF_CORRECT,WRONG
     }
-    public record CheckedResult(double score,double maxScore,CheckedResultType checkedResultType) {}
+    public record CheckedResult(float score,float maxScore,CheckedResultType checkedResultType) {}
     protected abstract void initFromSource(QuestionType question,Source source);
     public abstract CheckedResult check();
     public abstract QuestionType getSource();
